+++ components/tag/index.tsx
@@ -4,6 +4,8 @@ import Animate from 'rc-animate';
 import Icon from '../icon';
 import classNames from 'classnames';
 import splitObject from '../_util/splitObject';
+import omit from 'object.omit';
+
 export default class Tag extends React.Component {
   static defaultProps = {
     prefixCls: 'ant-tag',
@@ -46,8 +48,10 @@ export default class Tag extends React.Component {
   render() {
     const [{
       prefixCls, closable, color, className, children
-    },restProps] = splitObject(this.props,
-      ['prefixCls', 'closable', 'color','className','children']);
+    }, otherProps] = splitObject(
+      this.props,
+      ['prefixCls', 'closable', 'color','className','children']
+    );
     const close = closable ? <Icon type="cross" onClick={this.close} /> : '';
     const classString = classNames({
       [prefixCls]: true,
@@ -55,6 +59,11 @@ export default class Tag extends React.Component {
       [`${prefixCls}-close`]: this.state.closing,
       [className]: !!className,
     });
+    // fix https://fb.me/react-unknown-prop
+    const divProps = omit(otherProps, [
+      'onClose',
+      'afterClose',
+    ]);
     return (
       <Animate component=""
         showProp="data-show"
@@ -63,7 +72,7 @@ export default class Tag extends React.Component {
         onEnd={this.animationEnd}
       >
         {this.state.closed ? null : (
-          <div data-show={!this.state.closing} {...restProps} className={classString}>
+          <div data-show={!this.state.closing} {...divProps} className={classString}>
             <span className={`${prefixCls}-text`}>{children}</span>
             {close}
           </div>